requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# OAuth2 Integration
PyJWT==2.8.0
requests==2.31.0

# Utilities
Flask-CORS==4.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# PostgreSQL for Railway
psycopg2-binary==2.9.7

# File Storage
boto3==1.34.0
Werkzeug==3.0.1

# Testing
pytest==7.4.4
pytest-flask==1.3.0
Faker==37.5.3

# Optional (if Redis available)
# Flask-Limiter==3.5.0
# redis==5.0.1
